use native crypto for `keyRing.fromPrivate` and  `privateKey.fromSeed`
